Extract ShortLinkFromCsvRepository
